Type annotations for const variables

+++ src/TransferMatrix.jl
@@ -24,8 +24,8 @@ include("core.jl")
 include("dataio.jl")
 include("optics_functions.jl")
 
-const ε_0 = 8.8541878128e-12
-const μ_0 = 1.25663706212e-6
-const c_0 = 299792458
+const ε_0::Float64 = 8.8541878128e-12
+const μ_0::Float64 = 1.25663706212e-6
+const c_0::Float64 = 299792458
 
 end # module
